+++ transformer.py
@@ -109,6 +109,7 @@ class DataTransformer:
             "currency_rate": doc.currency_rate,
             "amount": doc.amount,
             "local_amount": doc.local_amount,
+            "unapplied_amount": doc.unapplied_amount,
             "cancelled": doc.cancelled,
             "details": doc.details,
             "company_category": self.entity.customer_category_map.get(doc.code, ""),